Refactoring of exceptions
