+++ src/Exception/InvalidArgumentException.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Cache\Exception;
 
-class InvalidArgumentException extends \RuntimeException implements \Psr\SimpleCache\InvalidArgumentException
+use Psr\SimpleCache\InvalidArgumentException as PsrInvalidArgumentException;
+
+final class InvalidArgumentException extends \InvalidArgumentException implements PsrInvalidArgumentException
 {
 }
