--- src/Exception/InvalidConfigException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Cache\Exception;
-
-final class InvalidConfigException extends CacheException
-{
-}
