+++ src/Exception/RemoveCacheException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Cache\Exception;
+
+final class RemoveCacheException extends CacheException
+{
+    public function __construct(string $key)
+    {
+        parent::__construct($key, 'Failed to delete the cache.');
+    }
+}
